+++ FlaskApp/core/config.py
@@ -2,3 +2,6 @@ MYSQL_USER = 'apostrophy_user'
 MYSQL_PASSWORD = ''
 MYSQL_DB = 'apostrophy'
 MYSQL_HOST = 'localhost'
+
+SQLALCHEMY_DATABASE_URI = 'mysql://apostrophy_user:@localhost/apostrophy'
+SQLALCHEMY_TRACK_MODIFICATIONS = False
