+++ FlaskApp/core/routes/__init__.py
@@ -1,0 +1,1 @@
+# Howdy! This is intentionally blank
