+++ FlaskApp/requirements.txt
@@ -1,3 +1,4 @@
+sqlalchemy
 werkzeug==0.14.1
 Flask==0.12.2
 Flask-Assets==0.12
